+++ commons/src/main/java/com/karuslabs/commons/command/aot/lexers/CommandLexer.java
@@ -44,19 +44,19 @@ public class CommandLexer implements Lexer {
     
     
     @Override
-    public List<Token> lex(Environment environment, Element location, String value, String context) {
+    public List<Token> lex(Environment environment, Element location, String value) {
         if (value.isBlank()) {
-            environment.error(location, "Command cannot be blank");
+            environment.error(location, "Command should not be blank");
             return EMPTY_LIST;
         }
         
         var tokens = new ArrayList<Token>();
         for (var command : value.split("\\s+")) {
             if (command.startsWith("<")) {
-                tokens.addAll(argument.lex(environment, location, command, context));
+                tokens.addAll(argument.lex(environment, location, command));
                 
             } else {
-                tokens.addAll(literal.lex(environment, location, command, context));
+                tokens.addAll(literal.lex(environment, location, command));
             }
         }