+++ commons/src/main/java/com/karuslabs/commons/command/aot/parsers/Parser.java
@@ -29,7 +29,7 @@ import com.karuslabs.commons.command.aot.lexers.Lexer;
 import java.util.List;
 import javax.lang.model.element.Element;
 
-import static com.karuslabs.commons.command.aot.Messages.reason;
+import static com.karuslabs.commons.command.aot.Messages.format;
 
 
 public abstract class Parser {
@@ -54,7 +54,7 @@ public abstract class Parser {
         
         var token = tokens.get(0);
         if (token.type == Type.ARGUMENT) {
-            environment.error(token.location, reason("Commands cannot start with arguments", token));
+            environment.error(token.location, format(token, "is at an invalid position", "a command should not start with an argument"));
             return false;
         }